requirements.txt
autopep8==1.5.3
flake8==3.8.2
PyQt5==5.15.0
PyQt5-sip==12.8.0
python-chess==0.31.2
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.2.3
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
